+++ app/src/main/java/com/henrystudio/moneymanager/data/model/TransactionGroup.kt
@@ -3,7 +3,7 @@ package com.henrystudio.moneymanager.data.model
 data class TransactionGroup(
     val id: Int = 0,
     val date: Long,
-    val income: Double,
-    val expense: Double,
+    val income: Long,
+    val expense: Long,
     val transactions: List<Transaction>
 )
