+++ app/src/main/java/com/henrystudio/moneymanager/presentation/bottomNavigation/dailyNavigate/DailyNavigateUiState.kt
@@ -4,12 +4,12 @@ import java.time.LocalDate
 
 data class DailyNavigateUiState(
     val monthLabel: String = "",
-    val incomeSum: Double = 0.0,
-    val expenseSum: Double = 0.0,
-    val totalSum: Double = 0.0,
+    val incomeSum: Long = 0L,
+    val expenseSum: Long = 0L,
+    val totalSum: Long = 0L,
     val selectionMode: Boolean = false,
     val selectedCount: Int = 0,
-    val selectedTotal: Double = 0.0,
+    val selectedTotal: Long = 0L,
     val currentTabPosition: Int = 0,
     val selectedDate: LocalDate = LocalDate.now(),
     val isYearlyView: Boolean = false
