+++ app/src/main/java/com/henrystudio/moneymanager/presentation/model/CategoryStat.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 data class CategoryStat(
     val name: String,
     val percent: Float,
-    val amount: Float,
+    val amount: Long,
     val color: Int
 ) : Serializable
