+++ app/src/main/java/com/henrystudio/moneymanager/presentation/model/Note.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 data class Note(
  val note: String,
  val count: Int,
- val amount: Double
+ val amount: Long
 ) : Serializable
