+++ app/src/main/java/com/henrystudio/moneymanager/presentation/monthly/MonthlyData.kt
@@ -6,9 +6,9 @@ data class MonthlyData(
     val monthName: String,
     val monthStart: LocalDate,
     val dateRange: String,
-    val income: Double,
-    val expense: Double,
-    val total: Double,
+    val income: Long,
+    val expense: Long,
+    val total: Long,
     val weeks: List<WeeklyData>,        // Danh sách các tuần
     var isExpanded: Boolean = false     // Trạng thái mở rộng
 )
