+++ app/src/main/java/com/henrystudio/moneymanager/presentation/monthly/WeeklyData.kt
@@ -5,8 +5,8 @@ import java.time.LocalDate
 data class WeeklyData(
     val weekStart: LocalDate,
     val weekRange: String,    // ví dụ: "11-05 ~ 17-05"
-    val income: Double,
-    val expense: Double,
-    val total: Double
+    val income: Long,
+    val expense: Long,
+    val total: Long
 )
