+++ app/src/main/java/com/henrystudio/moneymanager/presentation/viewmodel/StatisticAccountViewModel.kt
@@ -109,7 +109,7 @@ class StatisticAccountViewModel @Inject constructor() : ViewModel() {
             val amount = list.sumOf { it.amount }
             CategoryStat(
                 name = name,
-                amount = amount.toFloat(),
+                amount = amount,
                 percent = (amount / totalAmount * 100).toFloat(),
                 color = colors[index % colors.size]
             )
